+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## v3.0.0
+
+* Split `DistributedJob` in `DistributedJob::Client` and `DistributedJob::Job`
+* Add native namespace support and drop support for `Redis::Namespace`
+
 ## v2.0.0
 
 * `#push_each` no longer returns an enum when no block is given